LayoutTests/webanimations/animation-interface-effect-property.html

<script src="../resources/js-test-pre.js"></script>
<script>

description("Check the behavior of the effect property on Animation objects.");

debug("The animation.effect property should be null by default.");
const animation = new Animation;
shouldBeNull("animation.effect");

debug("");
debug("The animation.effect property can be set to a KeyframeEffect object.");
const effectA = new KeyframeEffect(document.body);
animation.effect = effectA;
shouldBe("animation.effect", "effectA");

debug("");
debug("The animation.effect property can be set to another KeyframeEffect object.");
const effectB = new KeyframeEffect(document.body);
animation.effect = effectB;
shouldBe("animation.effect", "effectB");

debug("");
debug("The animation.effect property can be set back to null.");
animation.effect = null;
shouldBeNull("animation.effect");

debug("");
debug("The animation.effect property cannot be set to something other than null or an AnimationEffect object.");
shouldThrow("animation.effect = 0");
shouldThrow("animation.effect = []");
shouldThrow("animation.effect = {}");
shouldThrow("animation.effect = false");
shouldThrow("animation.effect = 'effect'");

debug("");

</script>
<script src="../resources/js-test-post.js"></script>
