LayoutTests/webanimations/animation-interface-effect-property-expected.txt

Check the behavior of the effect property on Animation objects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


The animation.effect property should be null by default.
PASS animation.effect is null

The animation.effect property can be set to a KeyframeEffect object.
PASS animation.effect is effectA

The animation.effect property can be set to another KeyframeEffect object.
PASS animation.effect is effectB

The animation.effect property can be set back to null.
PASS animation.effect is null

The animation.effect property cannot be set to something other than null or an AnimationEffect object.
PASS animation.effect = 0 threw exception TypeError: The Animation.effect attribute must be an instance of AnimationEffect.
PASS animation.effect = [] threw exception TypeError: The Animation.effect attribute must be an instance of AnimationEffect.
PASS animation.effect = {} threw exception TypeError: The Animation.effect attribute must be an instance of AnimationEffect.
PASS animation.effect = false threw exception TypeError: The Animation.effect attribute must be an instance of AnimationEffect.
PASS animation.effect = 'effect' threw exception TypeError: The Animation.effect attribute must be an instance of AnimationEffect.

PASS successfullyParsed is true

TEST COMPLETE
